requirements.txt
alabaster==0.7.12
appdirs==1.4.4
astunparse==1.6.3
Babel==2.9.1
beautifulsoup4==4.9.3
black==21.5b1
certifi==2020.12.5
chardet==4.0.0
click==8.0.1
docutils==0.16
flake8==3.9.2
gevent==21.1.2
greenlet==1.0.0
idna==2.10
imagesize==1.2.0
importlib-metadata==4.3.0
Jinja2==3.0.1
loguru==0.5.3
MarkupSafe==2.0.1
mccabe==0.6.1
mypy-extensions==0.4.3
packaging==20.9
pathspec==0.8.1
pdoc==7.0.3
prettytable==2.1.0
pycodestyle==2.7.0
pydata-sphinx-theme==0.6.3
pyeventbus3==0.10
pyflakes==2.3.1
Pygments==2.9.0
pyparsing==2.4.7
pytz==2021.1
PyYAML==5.4.1
regex==2021.4.4
requests==2.25.1
six==1.16.0
snowballstemmer==2.1.0
soupsieve==2.2.1
Sphinx==3.5.4
sphinx-book-theme==0.1.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
tabulate==0.8.9
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.10.0.0
urllib3==1.26.5
wcwidth==0.2.5
zipp==3.4.1
zope.event==4.5.0
zope.interface==5.3.0
python-socketio==5.4.0
numpy~=1.21.1
graphviz~=0.17
pyzmq~=22.2.1